% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_counters.R
\name{CalculateDatesInMonths}
\alias{CalculateDatesInMonths}
\title{Calculate number of dates in months}
\usage{
CalculateDatesInMonths(dates)
}
\arguments{
\item{dates}{vector of dates}
}
\value{
data frame with columns: year (integer), month(integer), date_count (integer)
}
\description{
Calculates number of dates in every month in the given dates vector.
Missing values are omitted. Duplicated dates are counted multiple times!
}
\examples{
CalculateDatesInMonths(seq.Date(as.Date("2018-04-01"), as.Date("2018-06-19"), by = "1 day"))

}
