% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_counters.R
\name{CalculateDatesInYears}
\alias{CalculateDatesInYears}
\title{Calculate number of dates in years}
\usage{
CalculateDatesInYears(dates)
}
\arguments{
\item{dates}{vector of dates}
}
\value{
data frame with columns: year (integer), date_count (integer)
}
\description{
Calculates number of dates in every year in the given dates vector.
Missing values are omitted. Duplicated dates are counted multiple times!
}
\examples{
CalculateDatesInYears(seq.Date(as.Date("2018-04-01"), as.Date("2018-06-19"), by = "1 day"))

}
