% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hours_counters.R
\name{CalculateHoursInMonths}
\alias{CalculateHoursInMonths}
\title{Calculate number of hours in months}
\usage{
CalculateHoursInMonths(date_start, date_end)
}
\arguments{
\item{date_start}{beginning of the period, Date or object coercible to date}

\item{date_end}{end of the period, Date or object coercible to date}
}
\value{
data frame with columns: year (integer), month (integer), hour_count (integer)
}
\description{
Calculates number of dates in every month between (and including) the given dates.
Missing values are omitted.
}
\examples{
CalculateHoursInMonths("2018-01-01", "2018-12-31")

}
