% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hours_counters.R
\name{CalculateHoursInYears}
\alias{CalculateHoursInYears}
\title{Calculate number of hours in years}
\usage{
CalculateHoursInYears(date_start, date_end)
}
\arguments{
\item{date_start}{beginning of the period, Date or object coercible to date}

\item{date_end}{end of the period, Date or object coercible to date}
}
\value{
data frame with columns: year (integer), hour_count (integer)
}
\description{
Calculates number of dates in every year between (and including) the given dates.
Missing values are omitted.
}
\examples{
CalculateHoursInYears("2016-01-01", "2018-05-31")

}
