% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{Holidays}
\alias{Holidays}
\title{Find Polish holidays}
\usage{
Holidays(date_start, date_end = date_start, include_weekends = TRUE)
}
\arguments{
\item{date_start}{beginning of the period, Date or object coercible to date}

\item{date_end}{end of the period, Date or object coercible to date}

\item{include_weekends}{logical, whether or not weekends should be included}
}
\value{
vector of dates with holidays and weekends
}
\description{
Returns vector of all Polish holidays and weekends between (and including) th given dates
}
\examples{
Holidays("2018-03-28", "2018-04-03")
Holidays("2018-01-01", "2018-12-31", include_weekends = FALSE)

}
