% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{IsHoliday}
\alias{IsHoliday}
\title{Check for Polish holidays}
\usage{
IsHoliday(dates, include_weekends = TRUE)
}
\arguments{
\item{dates}{vector of Dates of object coercible by as.Date}

\item{include_weekends}{logical, whether or not weekends should be included}
}
\value{
logical vector of the same length as dates, TRUE when given date is
  holiday (or weekend), FALSE otherwise. Some holidays could fall on weekends.
}
\description{
Checks whether given dates are weekends or holidays
}
\examples{
IsHoliday("2020-04-12")
IsHoliday("2015-12-31")

}
