% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{MonthCounter}
\alias{MonthCounter}
\title{Calculates counts of every month for the given date-time vector}
\usage{
MonthCounter(date_times)
}
\arguments{
\item{date_times}{date-time object}
}
\value{
data frame with columns: year (integer), month (integer), count (integer)
}
\description{
Calculates counts of every month for the given date-time vector
}
