% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counters.R
\name{YearCounter}
\alias{YearCounter}
\title{Calculates counts of every year for the given date-time vector}
\usage{
YearCounter(date_times)
}
\arguments{
\item{date_times}{date-time object}
}
\value{
data frame with columns: year (integer), count (integer)
}
\description{
Calculates counts of every year for the given date-time vector
}
